update mesh_boxes function update bc in thermal problem

+++ fit/mesh/box.py
@@ -21,6 +21,7 @@ class Box:
         :param msh: Mesh object.
         :return: Index array.
         """
+
         xmin, xmax = self.x_range
         ymin, ymax = self.y_range
         zmin, zmax = self.z_range
@@ -43,7 +44,7 @@ def mesh_boxes(msh: Mesh, boxes: List[Box], mat: float) -> np.ndarray:
     :return: Material array of size ``(msh.np)``.
     """
 
-    mat = np.full(msh.np, mat)
+    mat = np.full(msh.np, mat, dtype=float)
     for box in boxes:
         mat[box.indices(msh)] = box.value
